Setup part of the flow for creating and editing a character.
Added a generic spinner adapter for ParseQueryAdapter as the default implementation doesn't work for spinners if you are targetting API > 19.
Changed PlayerCharacter to have references to CharacterClass and Race rather than just keeping their name. It seemed pretty close to impossible to set up a Spinner with the race/class if it didn't exist in the user's lists of races and classes.

Moved the create button for character from in the list fragment into the layouts for the character activity as it isn't related to the list's functionality.

Created a very basic/initial layout for fragment_character_upsert layout.

Added style for fragment layouts.
Added text styles to the theme.