Setup part of the flow for creating and editing an event.
Changed Event to have a reference to the PlayerCharacter in addition to just setting the ObjectId of the PlayerCharacter.

Moved the create button for event from in the list fragment into the layouts for the character activity as it isn't related to the list's functionality.

Created a very basic/initial layout for fragment_event_upsert layout.

Added states for the icon for adding an event.

Styling tweaks.