Implemented deleting characters.
Added another method to the BaseListFragment's Callbacks interface which should be called when the items for the list have been loaded.
Added an override of BaseListFragment.setSelection to highlight the new selection.

Added an orderByAscending on name to PlayerCharacter's getQuery.

Modified SimpleDialogFragment to allow setting the listener at the time of showing the dialog. This allows fragments to use it without requiring the hosting activity to implement the interface.

Created styles for primary and secondary buttons.