Implemented delete all characters and clear database.
Removed the hidden playercharacter id from the character list item view as it wasn't getting the character's objectId value after creation which was causing issues. This is now handled by getting the object for the list item view.

Enabled loading from local for CharacterClass and Race.

Changed logger to be a singleton as getting it via an activity was troublesome and causing crashes.